demo2/requirements.txt
click==6.7
Flask==0.12.2
greenlet==0.4.12
grpcio==1.4.0
grpcio-tools==1.4.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
msgpack-python==0.4.8
neovim==0.1.13
pip==9.0.1
protobuf==3.4.0
setuptools==28.8.0
six==1.10.0
Werkzeug==0.12.2
